- Added rpi qemu debug launcher - Added rpi qemu tasks - Added some debugging information - Changed set flag to singular - Changed different lpae layout to same as within short descriptor - Fixed ld_ttbcr_t structure - Removed mark of parameters as unused with void by using __unused - Changed v7 long create table to return correct address - Added logic to v7_long_prepare_temporary - Extended some todo point

+++ include/arch/arm/mm/virt/long.h
@@ -36,24 +36,24 @@
   #define LD_VIRTUAL_PAGE_INDEX( a ) ( ( a & 0x1FF000 ) >> 12 )
 
   typedef union __packed {
-    uint64_t raw;
+    uint32_t raw;
     struct {
-      uint64_t ttbr0_size : 3;
-      uint64_t sbz_0 : 4;
-      uint64_t ttbr0_disable_table_walk : 1;
-      uint64_t ttbr0_inner_cachability : 2;
-      uint64_t ttbr0_outer_cachability : 2;
-      uint64_t ttbr0_shareability : 2;
-      uint64_t sbz_1 : 2;
-      uint64_t ttbr1_size : 3;
-      uint64_t sbz_2 : 3;
-      uint64_t ttbr0_ttbr1_asid : 1;
-      uint64_t ttbr1_disable_table_walk : 1;
-      uint64_t ttbr1_inner_cachability : 2;
-      uint64_t ttbr1_outer_cachability : 2;
-      uint64_t ttbr1_shareability : 2;
-      uint64_t imp : 1;
-      uint64_t large_physical_address_extension : 1;
+      uint32_t ttbr0_size : 3;
+      uint32_t sbz_0 : 4;
+      uint32_t ttbr0_disable_table_walk : 1;
+      uint32_t ttbr0_inner_cachability : 2;
+      uint32_t ttbr0_outer_cachability : 2;
+      uint32_t ttbr0_shareability : 2;
+      uint32_t sbz_1 : 2;
+      uint32_t ttbr1_size : 3;
+      uint32_t sbz_2 : 3;
+      uint32_t ttbr0_ttbr1_asid : 1;
+      uint32_t ttbr1_disable_table_walk : 1;
+      uint32_t ttbr1_inner_cachability : 2;
+      uint32_t ttbr1_outer_cachability : 2;
+      uint32_t ttbr1_shareability : 2;
+      uint32_t imp : 1;
+      uint32_t large_physical_address_extension : 1;
     } data;
   } ld_ttbcr_t;
